Auf den rspack-Bundler umstellen
Meteors klassischer Bundler kennt das exports-Feld von package.json nicht
und liefert klassische Skripte statt ES-Module aus. Moderne npm-Pakete,
die import.meta in ihrem ESM-Build ausliefern, reissen damit das ganze
Bundle mit einem Parse-Fehler ab. rspack ist MDGs offizieller Nachfolger
und löst das.

  meteor add rspack
  meteor remove standard-minifier-js standard-minifier-css less

Die rspack-npm-Versionen bewusst nicht handverlesen: das Atmosphere-Paket
stellt beim ersten Lauf eine funktionierende Kombination bereit
(@meteorjs/rspack@2.0.1 mit @rspack/core@1.7.1). Nur less-loader/less
kommen von Hand dazu.

Zwei Dinge brauchten mehr als das Rezept:

Node-Builtins im Client-Bundle. imports/api/models/imgs.ts ist
isomorph -- imports/ui/Images.tsx braucht die Collection -- importierte
aber gm und fs auf oberster Ebene. Der klassische Bundler stopfte das
still per meteor-node-stubs zu und schickte den ImageMagick-Kleber in den
Browser; rspack bricht stattdessen ab. Die serverseitige Nachbearbeitung
liegt jetzt in server/imgs-processing.ts und hängt sich nachträglich ein.
Das ist zulässig, weil ostrio:files den Hook bei jedem Upload frisch von
der Instanz liest (server.js:1129) -- im Paketquelltext nachgesehen, nicht
angenommen. Der Code ist unverändert übernommen.

Stylesheets. Mit dem entfernten less-Package lädt Meteor .less nicht mehr
eager, rspack kennt kein Gegenstück -- rezepte.less wird darum explizit
aus client/main.tsx importiert. Reine .css-Dateien unter client/ zieht
Meteor weiterhin selbst ein: ein Import von fonts.css lud sie doppelt
(42 statt 21 @font-face-Regeln), er bleibt deshalb draussen.

Ausserdem:
- --port 0:3000 -> --port 3000; rspack leitet daraus seinen Dev-Server-Port
  ab und bekommt aus der host:port-Form NaN.
- eslint ignoriert die generierte rspack-Ausgabe, rspack.config.js darf
  CommonJS sein.
- stylesheets.d.ts deklariert die Seiteneffekt-Importe für tsc.
- Ein e2e-Test prüft, dass ein Theme-Token wirklich eine Farbe ergibt.
  Ohne ihn läuft die Suite auch auf einer völlig ungestylten App grün --
  gegengeprüft, indem der less-Import versuchsweise entfernt wurde.

Geprüft am gebauten Artefakt, nicht nur am Dev-Server: die
Selektor-Menge aus meteor build ist mit 132 zu 132 identisch zu vorher,
das Produktions-Bundle wurde ausgeführt und liefert dieselben berechneten
Styles. npm ci ohne Flags aus leerem node_modules läuft durch, und vom
Build-Werkzeug landet nichts im Server-Bundle.
