+++ app/.meteor/packages
@@ -8,8 +8,6 @@ meteor-base@1.5.2                   # Packages every Meteor app needs to have
 mongo@2.4.0                         # The database Meteor supports right now
 reactive-var@1.0.13                  # Reactive variable for tracker
 
-standard-minifier-css@1.10.1         # CSS minifier run for production mode
-standard-minifier-js@3.2.0          # JS minifier run for production mode
 ecmascript@0.18.1                    # Enable ECMAScript2015+ syntax in app code
 typescript@5.10.1
 shell-server@0.7.0                  # Server-side component of the `meteor shell` command
@@ -18,4 +16,4 @@ hot-module-replacement@0.6.0        # Update client in development without reloa
 static-html@1.5.0
 react-meteor-data
 ostrio:files
-less
+rspack
