+++ app/.meteor/versions
@@ -28,12 +28,9 @@ hot-code-push@1.0.5
 hot-module-replacement@0.6.0
 id-map@1.2.0
 inter-process-messaging@0.1.2
-less@4.1.1
 logging@1.3.6
 meteor@2.3.1
 meteor-base@1.5.2
-minifier-css@2.0.2
-minifier-js@3.2.1
 minimongo@2.2.0
 modern-browsers@0.2.3
 modules@0.20.3
@@ -55,12 +52,12 @@ reactive-var@1.0.13
 reload@1.3.2
 retry@1.1.1
 routepolicy@1.1.2
+rspack@1.1.0
 shell-server@0.7.0
 socket-stream-client@0.7.0
-standard-minifier-css@1.10.1
-standard-minifier-js@3.2.0
 static-html@1.5.0
 static-html-tools@1.0.0
+tools-core@1.1.0
 tracker@1.3.4
 typescript@5.10.1
 webapp@2.2.0
