+++ app/package.json
@@ -2,7 +2,7 @@
   "name": "rezeptee",
   "private": true,
   "scripts": {
-    "start": "meteor --exclude-archs web.browser.legacy --port 0:3000",
+    "start": "meteor --exclude-archs web.browser.legacy --port 3000",
     "visualize": "meteor --production --extra-packages bundle-visualizer",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
@@ -14,6 +14,7 @@
   "dependencies": {
     "@babel/runtime": "^8.0.0",
     "@datapunt/matomo-tracker-react": "^0.5.1",
+    "@swc/helpers": "^0.5.17",
     "classnames": "^2.3.1",
     "gm": "^1.23.1",
     "luxon": "^3.5.0",
@@ -38,7 +39,12 @@
   "devDependencies": {
     "@babel/preset-typescript": "^7.16.5",
     "@eslint/js": "^10.0.1",
+    "@meteorjs/rspack": "^2.0.1",
     "@playwright/test": "^1.52.0",
+    "@rsdoctor/rspack-plugin": "^1.5.7",
+    "@rspack/cli": "^1.7.1",
+    "@rspack/core": "^1.7.1",
+    "@rspack/plugin-react-refresh": "^1.4.3",
     "@types/gm": "^1.18.11",
     "@types/meteor": "^2.9.11",
     "@types/mocha": "^9.1.0",
@@ -49,7 +55,10 @@
     "eslint": "^10.8.0",
     "eslint-plugin-react-hooks": "^7.1.1",
     "globals": "^17.7.0",
+    "less": "^4.8.1",
+    "less-loader": "^13.0.0",
     "mocha": "^11.7.5",
+    "react-refresh": "^0.17.0",
     "tsx": "^4.21.0",
     "typescript": "^6.0.3",
     "typescript-eslint": "^8.65.0"
