+++ app/server/imgs-processing.ts
@@ -1,0 +1,60 @@
+import {FileRef} from "meteor/ostrio:files";
+import {Meteor} from "meteor/meteor";
+import fs from "fs";
+import gm from "gm";
+import {fsStorage, Imgs} from "/imports/api/models/imgs";
+
+const im = gm.subClass({ imageMagick: true });
+
+const bound = Meteor.bindEnvironment((callback) => {
+  return callback();
+});
+
+const createSizeVersion = (img: FileRef<any>, versionLabel: string, transform: (i: gm.State) => gm.State) => {
+  const versionPath = `${fsStorage}/${versionLabel}/${img._id}.avif`;
+
+  transform(im(img.path)).write(versionPath, (err) => {
+    fs.stat(versionPath, (err, stats) => {
+      bound(() => {
+        if (err) console.log(err)
+        return Imgs.update(img._id, {
+          $set: {
+            [`versions.${versionLabel}`]: {
+              path: versionPath,
+              size: stats.size,
+              type: 'image/avif',
+              name: img.name,
+            }
+          }
+        });
+      });
+    });
+  });
+}
+
+// Nachträglich statt via Konstruktor-Config: ostrio:files liest den Hook bei
+// jedem Upload frisch von der Instanz (`if (this.onAfterUpload)` in
+// server.js), und so bleiben gm und fs aus dem Client-Bundle heraus.
+Imgs.onAfterUpload = file => {
+  const image = im(file.path);
+  image.size((error, features) => {
+    bound(() => {
+      if (error) {
+        console.error('size not readable', error);
+        return;
+      }
+
+      // Update meta data if original image
+      Imgs.collection.update(file._id, {
+        $set: {
+          'meta.width': features.width,
+          'meta.height': features.height,
+          'versions.original.meta.width': features.width,
+          'versions.original.meta.height': features.height
+        }
+      });
+      createSizeVersion(file, 'thumbnail', i => i.quality(90).resize(300, undefined, '>').gravity('Center'));
+      createSizeVersion(file, 'full', i => i.quality(60).resize(1600, undefined, '>'));
+    })
+  }); // size + bound
+};
